+++ README.md
@@ -135,6 +135,7 @@
 - [x] MachineController 속 변수 및 메소드 클래스로 분리
 - [x] indent 1로 줄이기
 - [x] 링크드 리스트를 coinWallet 클래스로 분리
+- [x] 테스트 클래스 분리
 - [ ] 메소드가 한가지 일만 하도록 하기
 - [ ] InputController 속 메소드 반복 줄이고 구조화 하기
 - [ ] 상품 구매 시 인풋값 파트를 예외처리 및 재실행으로 변환
